Update Main.java
**DYNAMIC ID CARD IMPLEMENTATION**

//BEFORE//

Since each cards offers are linked to a specific ID, having the ID hardcoded leaded to scraping issues each time the lowest offer for a given card changed.

//NOW//

The script is looking for the first "articleRow" ID of the page (the cheapest product). Then it scrape the price and display it by concatenate this ID and the proper CSS Selector.